Plot integrations live in web2: series appends, f32 values, 10^6 rows
Task T2 of the new web UI (W6, #18; docs/ui-v2.md section 10): T2 done.

The web2 plot only saw a run when its command ended. Now the integrator
reports each stored row through a new XppUi hook, rows_stored() (called
after storind++ in integrate()'s loop, send_output and the equilibrium
range; headless and X11 keep the no-op default). ui_json.c's
j_rows_stored sends, at most every 100 ms and only to a client that asked
for series, {"ev":"series","op":"append","from":n,"rows":m,"columns":...}
with the columns of the last full series. It tracks how many of the
client's rows still equal storage (a run that starts again goes from row
0, Continue from the rows already there), sends a full series instead if
the curves changed mid-command, and always ends the command with the full
series as before, never an append after it.

`{"cmd":"data","enc":"f32"}` sends the columns as base64 of little-endian
float32 (core/series_enc.cpp, the new file in C++ with a C API; JSON
numbers stay the default and are formatted as buf_float did). web2 asks
for it: a million rows of three columns are 16 MB decoded with
Uint8Array.fromBase64 straight into the store's Float32Arrays, no copy.

The store appends in place: each column has a buffer that grows by
doubling, appends write past the rows an older state shows (a restart
gets new buffers), so states stay immutable and a run costs O(n) copying.
The chart keeps its uPlot instance and only sets new data.

uPlot's xy mode strokes every segment: 10^6 took 100-130 ms a draw. The
phase plane's lines and all point curves now use path builders from
plot/decimate.ts (pure, unit-tested): a segment that paints no device
pixel an earlier one has not painted is left out, so a limit cycle run
1500 times keeps ~1 900 vertices. The walk is O(n), so a long curve is
traced a slice per task (8 ms budget) while the last complete trace
stands in; appends continue the trace. Time plots keep uPlot's own
min/max-per-column line.

A long run's classic drawing ops (tens of MB) were parsed and thrown away
by web2 and made 60-90 ms tasks: /events?...&draw=0 (xpp_http.c) leaves
them out of that page's stream, and the transport skips any that come.

Tests: servercheck runs tools/models/live.ode (20 001 rows): several
appends, contiguous `from`, their rows the final series', nothing after
it, again from row 0 on a rerun, from 20001 for Continue, f32 equal to
JSON, nothing for an unsubscribed client. webcheck: a draw=0 stream has
no draw events. web2check: the live run's store and plot grow over the
appends and end equal to output.dat; tools/models/million.ode (1 000 001
rows): every draw under 50 ms during the run, wheel zooms of the phase
plane and the time plot with no draw over 50 ms and no long task
(__xpp.plot().drawMs, new __xpp.longTasks()). npm test covers decoding,
appends in place and the decimation.

Metrics: X11-free 99/121 -> 100/122, C++ 1/122 -> 2/123 sources;
warnings 0.

+++ web2/src/plot/model.ts
@@ -11,8 +11,8 @@ export interface CurveData {
   /** a line, or points of this radius */
   line: boolean;
   radius: number;
-  xs: Float64Array;
-  ys: Float64Array;
+  xs: Float32Array;
+  ys: Float32Array;
   /** storage row of xs[0] (rows before the largest shift are not plotted) */
   row0: number;
 }
@@ -23,12 +23,12 @@ export interface PlotModel {
   xLabel: string;
   yLabel: string;
   /** the time column, row by row, for the hover readout (null when not sent) */
-  t: Float64Array | null;
+  t: Float32Array | null;
 }
 
-const EMPTY = new Float64Array(0);
+const EMPTY = new Float32Array(0);
 
-function increasing(a: Float64Array): boolean {
+function increasing(a: Float32Array): boolean {
   for (let i = 1; i < a.length; i++) if (!(a[i] >= a[i - 1])) return false;
   return true;
 }
